Add support for multiple config files to bump
Added a new "files" property which can be set to an array of file
paths for files whose version property should be bumped.

The first file in the array is treated as the "primary", and all
the versions in the other files are synchronised to whatever the
primary's version is set to.

If "files" is set, the "file" property is ignored; however, if only
"file" is set, the version is bumped in that one file (maintained for
backwards-compatibility).